Log LRAUV application errors as errors. (#114)
In many cases we find "flaky" mission integration tests. Often this is because of components crashing on the LRAUV controller side. This often leads us down a very painful debugging journey. This patch proposes the following change: if there is an error reported by the LRAUV application, report it. The hope is that this will save us valuable time when debugging.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>