Check yaw rates instead of positions when yoyoing (retargeted #124 to main) (#133)
* bump depth expectations

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* Check yaw rates instead of positions when yoyoing

This  PR depends on #89, #96.

Recommended review order is #89 > #96 > this PR. If we want I can cherry pick this branch as the changes are independent of the underlying code. The merge order should be this PR > #96 > #89.

There is a slight change in the center of rotation of the yoyo mission with #96. This leads to test expectations failing. Rather than use position to check whether the vehicle is yoyoing, I think we should use @tfoote's suggestion and use yaw rate which makes the test independent of the center of rotation and hence more robust while at the same time ensuring that the vehicle actually moves in a circle.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* remove unused variable

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* Fix tolerances. :man-facepalming:

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* increase ramp up time

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>

* Fix bad merge and increase tolerance

Signed-off-by: Louise Poubel <louise@openrobotics.org>

Co-authored-by: Arjo Chakravarty <arjo@openrobotics.org>