Seperates the range bearing messages into two separate entities.
Previously we were using the same RangeBearing message for both external communication and internal communication. This leads to lots of confusion as we were using the same message in two different places. This PR also changes the public facing `LRAUVRangeBearingResponse` to stop using a vector to store the coordinates, but rather explicitly label the coordinates.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>