Adds a friendly warning if the python bindings for gazebo are not found (#260)
While installing gazebo on a new computer, I ran into trouble with python gazebo bindings.
I believe part of the issue is I have gazebo newly compiled from source. It's not great that
the repo does not compile at all if gz-python is not found, especially since gz-python is not
used by any core dependency. This PR adds a simple check and only runs the world generation
if the python-gazebo package is found. If the package is not found, users can still install the
simulator, however a warning is given warning users that the examples may not work.

I'm only checking for gz, but it might be worth checking for other python packages which may
or maynot come installed with an end user's work space.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>